code reorganization part 2
