+++ .eslintcache
@@ -1,1 +1,1 @@
-{"C:\\projects\\pst-server\\app.js":{"size":1766,"mtime":1511652689991,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\auth\\auth.js":{"size":0,"mtime":1511638281504,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\auth\\auth.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\auth\\local_auth.js":{"size":0,"mtime":1511638289643,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\auth\\local_auth.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\models\\user.js":{"size":0,"mtime":1511638318155,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\models\\user.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\base.js":{"size":877,"mtime":1511654993963,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\base.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\docs.js":{"size":370,"mtime":1511651161701,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\docs.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\index.js":{"size":268,"mtime":1511655115321,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\index.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\users.js":{"size":734,"mtime":1511572384612,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\users.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\tests\\index.test.js":{"size":282,"mtime":1511564211500,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\tests\\index.test.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\config\\app.js":{"size":541,"mtime":1511654996234,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\config\\app.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\config\\swagger.js":{"size":310,"mtime":1511651323037,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\config\\swagger.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\lib\\logger.js":{"size":0,"mtime":1511638407986,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\lib\\logger.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}}}
+{"C:\\projects\\pst-server\\app.js":{"size":1766,"mtime":1511652689991,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\auth\\auth.js":{"size":22,"mtime":1511655749318,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\auth\\auth.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\models\\user.js":{"size":22,"mtime":1511655767837,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\models\\user.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\base.js":{"size":877,"mtime":1511654993963,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\base.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\docs.js":{"size":370,"mtime":1511651161701,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\docs.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\index.js":{"size":268,"mtime":1511655115321,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\index.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\routes\\users.js":{"size":734,"mtime":1511655776052,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\routes\\users.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\api\\v1\\tests\\index.test.js":{"size":282,"mtime":1511564211500,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\api\\v1\\tests\\index.test.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\config\\app.js":{"size":544,"mtime":1511657667264,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\config\\app.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\config\\swagger.js":{"size":310,"mtime":1511651323037,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\config\\swagger.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}},"C:\\projects\\pst-server\\app\\lib\\logger.js":{"size":22,"mtime":1511655825619,"hashOfConfig":"1wexa8p","results":{"filePath":"C:\\projects\\pst-server\\app\\lib\\logger.js","messages":[],"errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0}}}
