+++ app/api/auth/auth.js
@@ -1,1 +1,4 @@
+// app/api/v1/auth/auth.js
+'use strict';
+
 module.exports = {};
