+++ app/api/auth/local.js
@@ -1,1 +1,4 @@
+// app/api/v1/auth/local.js
+'use strict';
+
 module.exports = {};
