+++ app/api/v1/models/user.js
@@ -1,1 +1,4 @@
+// app/api/v1/models/users.js
+'use strict';
+
 module.exports = {};
