+++ app/api/v1/routes/index.js
@@ -1,3 +1,6 @@
+// app/api/v1/routes/index.js
+'use strict';
+
 import indexRoute from './base';
 import usersRoute from './users';
 import docsRoute from './docs';
