+++ app/api/v1/routes/users.js
@@ -1,5 +1,36 @@
+// app/api/v1/routes/users.js
+'use strict';
+
 import express from 'express';
 const router = express.Router();
+
+// Handler: /users/
+router
+  .route('/')
+  .get((req, res, next) => {
+    res.send('should get users');
+  })
+  .post((req, res, next) => {
+    res.send('should create user');
+  });
+
+// Handler: /users/:user_id
+router
+  .route('/:userId')
+  .get((req, res, next) => {
+    res.send('should get user');
+  })
+  .put((req, res, next) => {
+    res.send('should update user');
+  })
+  .patch((req, res, next) => {
+    res.send('should patch user');
+  })
+  .delete((req, res, next) => {
+    res.send('should delete user');
+  });
+
+module.exports = router;
 
 /**
  * @swagger
@@ -31,9 +62,3 @@ const router = express.Router();
  *         schema:
  *           $ref: '#/definitions/Puppy'
  */
-
-router.get('/', function(req, res) {
-  res.send('respond with a list of users');
-});
-
-module.exports = router;
