+++ app/init/database.js
@@ -1,0 +1,4 @@
+// app/init/database.js
+'use strict';
+
+module.exports = {};
