+++ app/init/server.js
@@ -1,3 +1,6 @@
+// app/init/server.js
+'use strict';
+
 import express from 'express';
 import path from 'path';
 import middleware from '../middleware';
