+++ app/lib/logger.js
@@ -1,1 +1,4 @@
+// app/lib/logger.js
+'use strict';
+
 module.exports = {};
