+++ app/middleware/index.js
@@ -1,40 +1,14 @@
+// app/middleware/index.js
+'use strict';
+
 import morgan from 'morgan';
 import bodyParser from 'body-parser';
 import expressValidator from 'express-validator';
 import compression from 'compression';
 import helmet from 'helmet';
-import chalk from 'chalk';
 
 import appConfig from '../config/app';
-
-const padLeft = (str, len) => {
-  return len > str.length
-    ? new Array(len - str.length + 1).join(' ') + str
-    : str;
-};
-
-const padRight = (str, len) => {
-  return len > str.length
-    ? str + new Array(len - str.length + 1).join(' ')
-    : str;
-};
-
-const formatMorganString = (tokens, req, res) => {
-  var status = tokens.status(req, res);
-  var statusColor = status >= 500
-    ? 'red'
-    : status >= 400 ? 'yellow' : status >= 300 ? 'cyan' : 'green';
-
-  return (
-    chalk.reset(
-      padRight(tokens.method(req, res) + ' ' + tokens.url(req, res), 42)
-    ) +
-    ' ' +
-    chalk[statusColor](status) +
-    ' ' +
-    chalk.reset(padLeft(tokens['response-time'](req, res) + ' ms', 8))
-  );
-};
+import utils from '../utils/middleware';
 
 module.exports = () => {
   return app => {
@@ -43,7 +17,7 @@ module.exports = () => {
 
     // HTTP request morgan middleware
     let morganFormatter = appConfig[process.env.NODE_ENV].morgan === 'custom'
-      ? formatMorganString
+      ? utils.formatMorgan
       : appConfig[process.env.NODE_ENV].morgan;
 
     app.use(morgan(morganFormatter));
